+++ logging/bin/deploy_opensearch_content.sh
@@ -75,9 +75,10 @@ function set_retention_period {
 
     policy_name=$1
     retention_period_var=$2
+    pattern=$3
 
     # shellcheck disable=2145
-    log_debug "Function called: set_retention_perid ARGS: $@"
+    log_debug "Function called: set_retention_perid ARGS: $@ argCount [$#]}"
 
     retention_period=${!retention_period_var} # Retention Period (unit: days)
 
@@ -87,9 +88,10 @@ function set_retention_period {
 
     # confirm value is number
     if ! [[ $retention_period =~ $digits_re ]]; then
-        log_error "An invalid valid was provided for [$retention_period_var]; exiting."
+        log_error "An invalid value was provided for [$retention_period_var]; exiting."
         exit 1
     fi
+
 
     #Update retention period in json file prior to loading it
     sed -i'.bak' "s/\"min_index_age\": \"xxxRETENTION_PERIODxxx\"/\"min_index_age\": \"${retention_period}d\"/g" "$TMP_DIR"/"$policy_name".json
@@ -101,11 +103,23 @@ function set_retention_period {
     response=$(curl -s -o /dev/null -w "%{http_code}" -XPUT "$ism_api_url/policies/$policy_name" -H 'Content-Type: application/json' -d @"$TMP_DIR"/"$policy_name".json --user "$ES_ADMIN_USER":"$ES_ADMIN_PASSWD" --insecure)
     if [[ $response == 409 ]]; then
         log_info "The index management policy [$policy_name] already exist in OpenSearch; skipping load and using existing policy."
+        return 0
     elif [[ $response != 2* ]]; then
         log_error "There was an issue loading index management policy [$policy_name] into OpenSearch [$response]"
         exit 1
     else
         log_debug "Index management policy [$policy_name] loaded into OpenSearch [$response]"
+    fi
+
+    if [ -n "$pattern" ]; then
+        # Apply policy to existing indexes (will NOT override existing policy assignment) via API
+        response=$(curl -s -o /dev/null -w "%{http_code}" -XPOST "$ism_api_url/add/$pattern" -H 'Content-Type: application/json' -d '{"policy_id": "'"$policy_name"'"}' --user "$ES_ADMIN_USER":"$ES_ADMIN_PASSWD" --insecure)
+        if [[ $response != 2* ]]; then
+            log_error "There was an issue aplying the index management policy [$policy_name] to existing indexes matching the pattern [$pattern] [$response]"
+            exit 1
+        else
+            log_debug "Index management policy [$policy_name] applied to indexes matching the pattern [$pattern] that do NOT have an existing policy assigned [$response]"
+        fi
     fi
 }
 
@@ -174,7 +188,7 @@ function add_ism_template {
 set -e
 
 LOG_RETENTION_PERIOD="${LOG_RETENTION_PERIOD:-3}"
-set_retention_period viya_logs_idxmgmt_policy LOG_RETENTION_PERIOD
+set_retention_period viya_logs_idxmgmt_policy LOG_RETENTION_PERIOD 'viya_logs-*'
 add_ism_template "viya_logs_idxmgmt_policy" "viya_logs-*" 100
 
 # Create Ingest Pipeline to "burst" incoming log messages to separate indexes based on namespace
@@ -201,7 +215,7 @@ if [ "$OPENSHIFT_CLUSTER" == "true" ]; then
     # INFRASTRUCTURE LOGS
     # Handle "infrastructure" logs differently
     INFRA_LOG_RETENTION_PERIOD="${INFRA_LOG_RETENTION_PERIOD:-1}"
-    set_retention_period viya_infra_idxmgmt_policy INFRA_LOG_RETENTION_PERIOD
+    set_retention_period viya_infra_idxmgmt_policy INFRA_LOG_RETENTION_PERIOD 'viya_logs-openshift-*'
     add_ism_template "viya_infra_idxmgmt_policy" "viya_logs-openshift-*" 5
 
     # Link index management policy Index Template
@@ -221,7 +235,7 @@ OS_SECAUDIT_RETENTION_POLICY_ENABLE=${OS_SECAUDIT_RETENTION_POLICY_ENABLE:-true}
 if [ "$OS_SECAUDIT_RETENTION_POLICY_ENABLE" == "true" ]; then
 
     OS_SECAUDIT_RETENTION_PERIOD="${OS_SECAUDIT_RETENTION_PERIOD:-90}"
-    set_retention_period security-auditlog_idxmgmt_policy OS_SECAUDIT_RETENTION_PERIOD
+    set_retention_period security-auditlog_idxmgmt_policy OS_SECAUDIT_RETENTION_PERIOD 'security-auditlog-*'
 
     #NOTE: Since this policy was added post-ODFE, no need to call add_ism_template function
 
@@ -240,7 +254,7 @@ fi
 # ...index management policy automates the deletion of indexes after the specified time
 
 OPS_LOG_RETENTION_PERIOD="${OPS_LOG_RETENTION_PERIOD:-1}"
-set_retention_period viya_ops_idxmgmt_policy OPS_LOG_RETENTION_PERIOD
+set_retention_period viya_ops_idxmgmt_policy OPS_LOG_RETENTION_PERIOD 'viya_ops-*'
 add_ism_template "viya_ops_idxmgmt_policy" "viya_ops-*" 50
 
 # Load template